+++ verl_omni/custom_pipelines/qwen_image/diffusers_training_adapter.py
@@ -26,8 +26,8 @@ from tensordict import TensorDict
 from verl.utils import tensordict_utils as tu
 from verl.utils.device import get_device_name
 
+from verl_omni.custom_pipelines.base import DiffusionModelBase
 from verl_omni.custom_pipelines.schedulers import FlowMatchSDEDiscreteScheduler
-from verl_omni.models.diffusion_model import DiffusionModelBase
 from verl_omni.workers.config import DiffusionModelConfig
 
 from .common import QWEN_IMAGE_VAE_SCALE_FACTOR, apply_true_cfg, build_img_shapes
