--- verl_omni/models/diffusion_model/__init__.py
@@ -1,25 +1,0 @@
-# Copyright 2026 Bytedance Ltd. and/or its affiliates
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-from .base import DiffusionModelBase, VllmOmniPipelineBase
-from .utils import build_scheduler, forward_and_sample_previous_step, prepare_model_inputs, set_timesteps
-
-__all__ = [
-    "DiffusionModelBase",
-    "VllmOmniPipelineBase",
-    "build_scheduler",
-    "set_timesteps",
-    "prepare_model_inputs",
-    "forward_and_sample_previous_step",
-]
