+++ verl_omni/workers/engine/fsdp/diffusers_impl.py
@@ -60,7 +60,7 @@ from verl.workers.engine.base import BaseEngine, BaseEngineCtx, EngineRegistry
 from verl.workers.engine.fsdp.utils import create_device_mesh, get_sharding_strategy
 from verl.workers.engine.utils import enable_full_determinism, prepare_micro_batches
 
-from verl_omni.models.diffusion_model import build_scheduler, forward_and_sample_previous_step, prepare_model_inputs
+from verl_omni.custom_pipelines.utils import build_scheduler, forward_and_sample_previous_step, prepare_model_inputs
 from verl_omni.workers.config import DiffusionModelConfig
 
 logger = logging.getLogger(__file__)
