+++ verl_omni/workers/rollout/vllm_rollout/vllm_omni_async_server.py
@@ -37,7 +37,7 @@ from vllm_omni.inputs.data import OmniCustomPrompt, OmniDiffusionSamplingParams
 from vllm_omni.lora.request import LoRARequest
 from vllm_omni.outputs import OmniRequestOutput
 
-from verl_omni.models.diffusion_model import VllmOmniPipelineBase
+from verl_omni.custom_pipelines.base import VllmOmniPipelineBase
 from verl_omni.workers.config import DiffusionModelConfig, DiffusionRolloutConfig
 from verl_omni.workers.rollout.replica import DiffusionOutput
